refactor forecast to helpers
